requirements.txt
annotated-types==0.7.0
anyio==4.6.0
click==8.1.7
fastapi==0.115.0
h11==0.14.0
idna==3.10
orjson==3.10.7
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
simplejson==3.19.3
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.1
